tweaks

+++ create-app.sh
@@ -1,19 +1,22 @@
 #!/bin/sh
 
-# 1. run your program in eclipse
-# 2. run this script: ./create-app.sh [main-class] [path-to-eclipse-bin-folder]
-# 3. run the AppRunner/AppRunner.pde sketch in this directory from Processing
-# 4. choose file->'export application' from Processing menu
-# 5. test the exported application
+# Assumptions: running from OSX, with processing-java installed
 #
-# Assumptions: running from OSX
+# 1. run your program in eclipse
+# 2. then run this script
+# 3. test the exported application(s)
+#
+
+if [ $# -lt 2 ]; then
+  echo 1>&2 "Usage: ./create-app.sh [main-class] [path-to-eclipse-project] (optional-path-to-jars-dir)"
+  exit 2
+fi
 
 MAIN_CLASS=${1}
 PROJECT_HOME=${2}
 PROJECT_BIN=${PROJECT_HOME}/bin
 JARS_DIR=${3:-${PROJECT_HOME}/lib}
 DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
-#CORE_JAR=${3:-/Applications/Processing.app/Contents/Java/core.jar}
 
 # set things up
 rm -rf tmp
@@ -29,7 +32,7 @@ cat template.pde | sed -e "s/%MAIN_CLASS%/${MAIN_CLASS}/" > AppRunner/Apprunner.
 cp -r ${PROJECT_BIN}/* tmp/
 
 # and create a jar in code folder
-cd tmp && jar cvf ../AppRunner/code/app.jar * && cd -
+cd tmp && jar cf ../AppRunner/code/app.jar * && cd -
 
 # copy rest of jars from eclipse folder
 cp -r $JARS_DIR/*.jar AppRunner/code/
@@ -40,8 +43,12 @@ rm -rf  AppRunner/application.* # only exists if run repeatedly
 rm -rf /Users/$USER/.Trash/application.* # remove from trash as will cause an error
 
 # lets do the export
-processing-java --sketch=${DIR}/AppRunner --force --export
+processing-java --sketch=${DIR}/AppRunner --force --export > AppRunner.log 2>&1
+
+# now, copy any natives from eclipse folder
+cp -r $PROJECT_HOME/*.jnilib AppRunner/application.macosx/AppRunner.app/Contents/Java
 
 # remove our garbage and open the folder
 rm -rf tmp
+echo done
 open ./AppRunner
